requirements.txt
Django==1.5.4
South==0.8.2
django-embed-video==0.5
django-tastypie==0.10.0
python-dateutil==2.1
requests==2.0.0
six==1.4.1
wsgiref==0.1.2
